Fix build for Android with android-clang
kernel/qnetworkinterface_linux.cpp:204:17: error: comparison of integers of different signs: '__u32' (aka 'unsigned int') and 'qsizetype' (aka 'int') [-Werror,-Wsign-compare]
            if (NLMSG_OK(hdr, len))
                ^~~~~~~~~~~~~~~~~~

This amends 09cb23f342fd2eae7ca85a99fa0a10b7ab103443.

+++ src/network/kernel/qnetworkinterface_linux.cpp
@@ -201,7 +201,7 @@ template <typename Lambda> struct ProcessNetlinkRequest
             }
 
 #ifndef QT_NO_DEBUG
-            if (NLMSG_OK(hdr, len))
+            if (NLMSG_OK(hdr, quint32(len)))
                 qWarning("QNetworkInterface/AF_NETLINK: received unknown packet type (%d) or too short (%u)",
                          hdr->nlmsg_type, hdr->nlmsg_len);
             else
